pyproject.toml
[tool.poetry]
name = "tic-tac-toe"
version = "1.2.0"
description = ""
authors = ["Nyaaa <1457047+Nyaaa@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "tic_tac_toe"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.4"
termcolor = "2.0.1"
python-dotenv = "^0.21.0"
tabulate = "^0.9.0"
pytelegrambotapi = "^4.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
